Fix Orders table styles
